# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Matrix Attual Core API
  version: 0.4.0
  description: >-
    Matrix Attual API for consent-gated AttualPlay analytics, in-app topic
    personalization and the M4 explicit ATTUAL ONE identity bridge. M4 does
    not enable marketing or autonomous external messaging.
servers:
  - url: https://matrix-attual-core.vercel.app
    description: Matrix production API
paths:
  /health:
    get:
      operationId: health
      responses:
        '200':
          description: Process healthy
          content:
            application/json:
              schema:
                type: object
                required: [status, service, version, milestone]
                properties:
                  status: { type: string, const: ok }
                  service: { type: string, const: matrix-event-api }
                  version: { type: string, const: 0.4.0 }
                  milestone: { type: string, const: M4 }
  /ready:
    get:
      operationId: ready
      responses:
        '200': { description: Matrix database ready }
        '503': { description: Matrix database not ready }
  /v1/event-types:
    get:
      operationId: listEventTypes
      responses:
        '200': { description: Event Catalog v1 types accepted by the API }
  /v1/events:
    post:
      operationId: ingestEvent
      security:
        - matrixClient: []
          matrixPublishableKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventEnvelope' }
      responses:
        '200':
          description: Accepted or idempotent duplicate
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestResult' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
  /v1/events/batch:
    post:
      operationId: ingestEventBatch
      security:
        - matrixClient: []
          matrixPublishableKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [events]
              properties:
                events:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items: { $ref: '#/components/schemas/EventEnvelope' }
      responses:
        '200': { description: Batch processed }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
  /v1/recommendations/query:
    post:
      operationId: queryRecommendation
      description: >-
        Returns an eligible in-app topic recommendation only after explicit
        analytics, personalization and adult confirmation. Identified queries
        are additionally gated by the current server-side consent ledger.
      security:
        - matrixClient: []
          matrixPublishableKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecommendationQuery' }
      responses:
        '200':
          description: Eligible recommendation or null
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecommendationResponse' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /v1/internal/identity/bridge:
    post:
      operationId: issueIdentityBridge
      description: >-
        Server-only endpoint used by ATTUAL ONE to issue a temporary one-time
        bridge code. The external account reference is an opaque UUID; name,
        email, phone and CPF are not accepted by this contract.
      security:
        - attualOneServer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IdentityBridgeIssue' }
      responses:
        '201': { description: One-time bridge code issued }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /v1/identity/link:
    post:
      operationId: consumeIdentityBridge
      description: >-
        Explicitly links the current AttualPlay anonymous profile to the Matrix
        person represented by a valid one-time bridge code and creates a
        revocable opaque person-session token.
      security:
        - matrixClient: []
          matrixPublishableKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IdentityLinkRequest' }
      responses:
        '200': { description: Identity linked and person session issued }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/consents:
    post:
      operationId: syncPersonConsent
      description: Persists current identified analytics/personalization consent in the Matrix server ledger.
      security:
        - matrixClient: []
          matrixPublishableKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PersonConsentRequest' }
      responses:
        '200': { description: Server-side consent synchronized }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /v1/identity/unlink:
    post:
      operationId: unlinkIdentity
      description: Revokes the active Matrix person session and explicit identity link for the project.
      security:
        - matrixClient: []
          matrixPublishableKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IdentityUnlinkRequest' }
      responses:
        '200': { description: Identity session/link revoked }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /v1/internal/m4/dispatch:
    post:
      operationId: dispatchM4Signals
      description: Server-only governed dispatcher for eligible ATTUAL ONE qualified-interest signals.
      security:
        - m4InternalServer: []
      responses:
        '200': { description: Dispatch cycle completed; marketing and n8n external actions remain disabled }
        '401': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /v1/internal/m4/status:
    get:
      operationId: m4Status
      description: Server-only M4 observability snapshot.
      security:
        - m4InternalServer: []
      responses:
        '200': { description: M4 observability snapshot }
        '401': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    matrixClient:
      type: apiKey
      in: header
      name: X-Matrix-Client
      description: Publishable producer identifier, currently attualplay.
    matrixPublishableKey:
      type: apiKey
      in: header
      name: X-Matrix-Key
      description: Low-authority publishable Matrix key; never a database administrative key.
    attualOneServer:
      type: http
      scheme: bearer
      description: Server-only Matrix/ATTUAL ONE integration credential.
    m4InternalServer:
      type: http
      scheme: bearer
      description: Server-only credential for M4 dispatcher and observability routes.
  schemas:
    ConsentSnapshot:
      type: object
      additionalProperties: false
      properties:
        essential: { type: boolean, default: true }
        analytics: { type: boolean }
        personalization: { type: boolean }
        marketing: { type: boolean }
        adult_confirmed: { type: boolean }
        policy_version: { type: string, maxLength: 80 }
    M4Consent:
      type: object
      additionalProperties: false
      required: [analytics, personalization, adult_confirmed, marketing, policy_version]
      properties:
        analytics: { type: boolean }
        personalization: { type: boolean }
        adult_confirmed: { type: boolean }
        marketing: { type: boolean, const: false }
        policy_version: { type: string, minLength: 3, maxLength: 80 }
    EventEnvelope:
      type: object
      additionalProperties: false
      required: [event_id, event_type, occurred_at, project_key, idempotency_key]
      anyOf:
        - required: [anonymous_id]
        - required: [person_token]
      properties:
        event_id: { type: string, format: uuid }
        event_type: { type: string, minLength: 3, maxLength: 128 }
        occurred_at: { type: string, format: date-time }
        project_key: { type: string, minLength: 2, maxLength: 100 }
        anonymous_id: { type: [string, 'null'], minLength: 3, maxLength: 200 }
        person_token: { type: [string, 'null'], minLength: 8, maxLength: 512 }
        session_id: { type: [string, 'null'], maxLength: 200 }
        object:
          type: [object, 'null']
          additionalProperties: false
          required: [type, id]
          properties:
            type: { type: string, maxLength: 80 }
            id: { type: string, maxLength: 200 }
        properties: { type: object, default: {} }
        context: { type: object, default: {} }
        consent: { $ref: '#/components/schemas/ConsentSnapshot' }
        idempotency_key: { type: string, minLength: 8, maxLength: 200 }
    IngestResult:
      type: object
      required: [accepted, duplicate, event_id, correlation_id]
      properties:
        accepted: { type: boolean, const: true }
        duplicate: { type: boolean }
        event_id: { type: string, format: uuid }
        correlation_id: { type: string, format: uuid }
    RecommendationQuery:
      type: object
      additionalProperties: false
      required: [project_key, consent]
      anyOf:
        - required: [anonymous_id]
        - required: [person_token]
      properties:
        project_key: { type: string, minLength: 2, maxLength: 100 }
        anonymous_id: { type: string, minLength: 3, maxLength: 200 }
        person_token: { type: string, minLength: 16, maxLength: 512 }
        consent:
          type: object
          additionalProperties: false
          required: [analytics, personalization, adult_confirmed, policy_version]
          properties:
            analytics: { type: boolean, const: true }
            personalization: { type: boolean, const: true }
            adult_confirmed: { type: boolean, const: true }
            marketing: { type: boolean }
            policy_version: { type: string, minLength: 3, maxLength: 80 }
    RecommendationResponse:
      type: object
      required: [recommendation, policy, correlation_id]
      properties:
        recommendation: { type: [object, 'null'] }
        policy:
          type: object
          properties:
            personalization_opt_in: { type: boolean, const: true }
            server_side_consent: { type: boolean }
            marketing_enabled: { type: boolean, const: false }
        correlation_id: { type: string, format: uuid }
    IdentityBridgeIssue:
      type: object
      additionalProperties: false
      required: [project_key, external_system, external_account_ref]
      properties:
        project_key: { type: string, minLength: 2, maxLength: 100 }
        external_system: { type: string, const: attual_one }
        external_account_ref: { type: string, format: uuid }
    IdentityLinkRequest:
      type: object
      additionalProperties: false
      required: [project_key, anonymous_id, bridge_code, consent]
      properties:
        project_key: { type: string }
        anonymous_id: { type: string, minLength: 3, maxLength: 200 }
        bridge_code: { type: string, minLength: 12, maxLength: 120 }
        consent: { $ref: '#/components/schemas/M4Consent' }
    PersonConsentRequest:
      type: object
      additionalProperties: false
      required: [project_key, person_token, consent]
      properties:
        project_key: { type: string }
        person_token: { type: string, minLength: 16, maxLength: 512 }
        consent: { $ref: '#/components/schemas/M4Consent' }
    IdentityUnlinkRequest:
      type: object
      additionalProperties: false
      required: [project_key, person_token]
      properties:
        project_key: { type: string }
        person_token: { type: string, minLength: 16, maxLength: 512 }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlation_id]
          properties:
            code: { type: string }
            message: { type: string }
            correlation_id: { type: string, format: uuid }
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }